Separate interface from dag and add reduction of variables to interface.
