+++ gettsim/tests/test_dag.py
@@ -1,7 +1,7 @@
 import pytest
 
-from gettsim.dag import fail_if_functions_and_user_columns_overlap
-from gettsim.dag import fail_if_user_columns_are_not_in_data
+from gettsim.interface import _fail_if_functions_and_user_columns_overlap
+from gettsim.interface import _fail_if_user_columns_are_not_in_data
 
 
 def test_fail_if_user_columns_are_not_in_data():
@@ -13,7 +13,7 @@ def test_fail_if_user_columns_are_not_in_data():
     ]
 
     with pytest.raises(ValueError):
-        fail_if_user_columns_are_not_in_data(data, user_columns)
+        _fail_if_user_columns_are_not_in_data(data, user_columns)
 
 
 @pytest.mark.parametrize("type_", ["user", "internal"])
@@ -21,4 +21,4 @@ def test_fail_if_user_functions_and_user_columns_overlap(type_):
     dict_ = {"a": None, "b": None, "c": None, "d": None}
 
     with pytest.raises(ValueError):
-        fail_if_functions_and_user_columns_overlap(dict_, dict_, type_, [])
+        _fail_if_functions_and_user_columns_overlap(dict_, dict_, type_, [])
