+++ gettsim/tests/test_kindergeld.py
@@ -6,7 +6,7 @@ import pytest
 from pandas.testing import assert_series_equal
 
 from gettsim.config import ROOT_DIR
-from gettsim.dag import compute_taxes_and_transfers
+from gettsim.interface import compute_taxes_and_transfers
 from gettsim.pre_processing.policy_for_date import get_policies_for_date
 from gettsim.tests.auxiliary import select_output_by_level
