+++ gettsim/tests/test_soz_vers.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pytest
 
 from gettsim.config import ROOT_DIR
-from gettsim.dag import compute_taxes_and_transfers
+from gettsim.interface import compute_taxes_and_transfers
 from gettsim.pre_processing.policy_for_date import get_policies_for_date
 
 INPUT_COLS = [
